change default export path + validate export directory + move IsChildOfKnownPrefab
