+++ Scripts/slocExporter/Serialization/Exporting/ExportCollector.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using slocExporter.Extensions;
-using slocExporter.Serialization.Exporting.Identifiers;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.SceneManagement;
@@ -50,19 +49,6 @@ namespace slocExporter.Serialization.Exporting
         public static bool ShouldInclude(this ExportPreset preset, GameObject gameObject)
             => (gameObject.activeSelf || preset.includeInactiveObjects)
                && (preset.traversePrefabs || !gameObject.IsChildOfAnyPrefab(true, out _));
-
-        public static bool IsChildOfKnownPrefab(this GameObject gameObject)
-        {
-            if (!gameObject.IsChildOfAnyPrefab(false, out var isRoot))
-                return false;
-            var root = isRoot ? gameObject : PrefabUtility.GetNearestPrefabInstanceRoot(gameObject);
-            if (!root.TryGetPrefabGuid(out var guid, false))
-                return false;
-            if (guid == CapybaraIdentifier.CapybaraGuid)
-                return true;
-            var guidString = guid.ToString();
-            return Identify.CameraGuids.ContainsKey(guidString) || Identify.StructureGuids.ContainsKey(guidString);
-        }
 
     }
